requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
pyyaml>=6.0.0
psutil>=5.9.0
googlesearch-python>=3.0.0
beautifulsoup4>=4.12.0